telegram_bot/requirements.txt
python-telegram-bot==21.9
asyncpg==0.30.0
paho-mqtt==1.6.1